+++ src/market/ticker.rs
@@ -1,11 +1,13 @@
+use crate::decimal::Decimal;
+
 #[derive(Debug, Clone)]
 pub struct PriceTick {
     pub symbol: String,
-    pub price: f64,
+    pub price: Decimal,
 }
 
 impl PriceTick {
-    pub fn new(symbol: &str, price: f64) -> Self {
+    pub fn new(symbol: &str, price: Decimal) -> Self {
         Self {
             symbol: symbol.to_string(),
             price,
